Avoid hang when restarting
When doing a restart of the service, if a service is down and
can't be connect()'d to, then the shutdown hangs waiting on
threads to exit.  Give the threads a little bit of time and
then just continue on.  This is just a workaround and not
upstream worthy but good to verify the theory with a customer

https://github.com/collectd/collectd/issues/315

original commit 14126858baf4f03cefe51dcbbcbe6c7105532c78