SNMP Agent plugin: Fix - not every OID is unregistered from snmp table
Problem occures when a metric has more than one OID assigned. Then,
during the table unregistration, plugin thinks that there are some OIDs
left in the table and keeps index keys also registered. It's because
plugin registers all the OIDs for the particular metric, but unregisters only
the first one which is wrong.

Change-Id: I6daae1ceeebf30b98f22075940b29936ee3d110d
Signed-off-by: Mozejko, MarcinX <marcinx.mozejko@intel.com>