chrony: Ignore late responses
If a response from chronyd was not received before the timeout (e.g. on a
busy system), it would wait in the receive queue and be processed as a
response of the next request. It would fail the sequence check and cause
the plugin to return with an error. The latest response would wait in
the receive queue and this would repeat forever.

Flush the receive queue before sending a new batch of requests to
prevent this situation from happening.