dpdk: fix dpdk runtime config file path
The config path was changed in DPDK by the commit 'adf1d86  eal: move
runtime config file to new location’. As a result, the dpdk plugins won't
be able to link to the DPDK primary application after this change.

This commit adds a compile time check for the DPDK verison being used and
sets the DPDK runtime config file location accordingly.

Suggested-by: Michael Luo <michael.luo@intel.com>
Signed-off-by: Kevin Laatz <kevin.laatz@intel.com>